% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_sales.R
\name{add_istanbul_part_city_sales}
\alias{add_istanbul_part_city_sales}
\title{Add istanbul part to city sales}
\usage{
add_istanbul_part_city_sales(city_sales)
}
\description{
This is harder than it sounds. City sales data provided by tuik doesn't
seperate İstanbul to continents. This is approximated using the
\code{town_sales} data. Since we know each part of İstanbul's towns,
a ratio is calculated and then used to seperate first_hand and second_hand
sales for each part of İstanbul.
}

