% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_city_sales.R
\docType{data}
\name{city_sales}
\alias{city_sales}
\title{City apartment sales of Turkey}
\format{A data.frame with 5 variables
\describe{
   \item{city}{name of the city}
   \item{sale_type}{type of the sale, first hand or second hand}
   \item{year}{year the measurement was taken}
   \item{month}{month the measurement was taken}
   \item{sale}{amount of the sale}
}}
\source{
explained extensively on \code{inst/extdata/README.md}
}
\usage{
city_sales
}
\description{
A dataset containing apartment sales of Turkey's cities
}
\keyword{datasets}

