% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_city_sales.R
\name{clean_city_sales}
\alias{clean_city_sales}
\title{Cleaning city sales data from TUIK}
\usage{
clean_city_sales(folder_path = system.file("extdata/city_sales", package =
  "artuik"))
}
\arguments{
\item{file_path}{path of the city sales data folder}
}
\description{
The process of obtaining raw data is explained in
\code{inst/extdata/README.md}. This function takes that raw data and makes
it tidy data.
}
\details{
Data generated from here is later put into \code{data} folder by
\code{devtools::use_data()}
}

