% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_city_sales.R, R/clean_town_population.R
\name{clean_town_population}
\alias{clean_town_population}
\title{Cleaning city sales data from TUIK}
\usage{
clean_town_population(file_path = system.file("extdata/town_population.csv",
  package = "artuik"))

clean_town_population(file_path = system.file("extdata/town_population.csv",
  package = "artuik"))
}
\arguments{
\item{file_path}{path of the town population raw data}

\item{file_path}{path of the town population raw data}
}
\description{
The process of obtaining raw data is explained in
\code{inst/extdata/README.md}. This function takes that raw data and makes
it tidy data.

The process of obtaining raw data is explained in
\code{inst/extdata/README.md}. This function takes that raw data and makes
it tidy data.
}
\details{
Data generated from here is later put into \code{data} folder by
\code{devtools::use_data()}

Data generated from here is later put into \code{data} folder by
\code{devtools::use_data()}
}

