% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_sales.R
\name{get_city_sales}
\alias{get_city_sales}
\title{Tuik city sales aggregated}
\usage{
get_city_sales(istanbul_part = FALSE, first_total_ratio = FALSE,
  start_date = NULL, end_date = NULL)
}
\arguments{
\item{istanbul_part}{whether istanbul continent part should be added to city.
If added, this will be an approximation based on the
\code{\link{town_sales}} data.}

\item{first_total_ratio}{if \code{TRUE}, first sales ratio to total sales
ratio is added with the argument name.}

\item{total}{whether the sales should be aggregated in total. default is
per month.}
}
\description{
City sales are per month. This function creates variations.
}

