% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/town_sales.R
\name{get_europe_istanbul_ratio}
\alias{get_europe_istanbul_ratio}
\title{Istanbul europe side home sales to total ratio}
\usage{
get_europe_istanbul_ratio(average = FALSE)
}
\arguments{
\item{average}{rather then returning the ratio per month, return one number
for the average ratio.}
}
\description{
\code{town_sales} data has the granularity to calculate this ratio.
Here the calcualted ratio is the total home sales. Normally \code{town_sales}
data is not complete because there is no data for every town, but for
İstanbul, it is complete enough.
}

