% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/town_population.R
\name{get_town_population}
\alias{get_town_population}
\title{Get town population data}
\usage{
get_town_population(istanbul_part = FALSE)
}
\arguments{
\item{istanbul_part}{Adds istanbul continent part to the city variable}
}
\description{
Getting the TUIK \code{town_population} data.
}

