% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/town_sales.R
\name{get_town_sales}
\alias{get_town_sales}
\title{Get town sales data}
\usage{
get_town_sales(istanbul_part = FALSE, first_sale = FALSE)
}
\arguments{
\item{istanbul_part}{Adds istanbul continent part to the city variable}

\item{first_sale}{approximates first hand house sales using the
\code{city_sales} data. This is only an approximation. This data is not
provided by TUIK for towns.}
}
\description{
Getting the TUIK \code{town_sales} data. Also creates some derivations with
function arguments.
}

