% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_utils.R
\name{get_tuik_month}
\alias{get_tuik_month}
\title{Getting month as numeric from TUIK data}
\usage{
get_tuik_month(month_vec)
}
\description{
TUIK gives months like this "01-January". This function is just for getting
c(1) out of that.
}

