% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuts.R
\docType{data}
\name{nuts_level}
\alias{nuts_level}
\title{Statistical region levels of Turkey}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 81 rows and 3 columns.}
\source{
http://ec.europa.eu/eurostat/en/web/products-manuals-and-guidelines/-/KS-RA-07-023
}
\usage{
nuts_level
}
\description{
Statistical region levels of Turkey
}
\keyword{datasets}

