% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_utils.R
\name{remove_last_NA_row}
\alias{remove_last_NA_row}
\title{Remove last row if its filled with \code{NA}'s}
\usage{
remove_last_NA_row(frame)
}
\description{
This is a comman case among CSV obtained from TUIK.
This function is used during the cleaning process.
}

