% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_utils.R
\name{split_city_town}
\alias{split_city_town}
\title{Split concatanated city and town data}
\usage{
split_city_town(frame)
}
\description{
TUIK data merges city and town data in this format: "Adana(Aladağ)".
Cleaner functions in this package puts them in a column called
\code{city_town}. This function seperates them and returns them in seperate
\code{city} and \code{town} varialables.
}

