% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_town_population.R
\docType{data}
\name{town_population}
\alias{town_population}
\title{Town population of Turkey}
\format{A data.frame with 4 variables
\describe{
   \item{city}{name of the city}
   \item{town}{name of the town}
   \item{year}{year the measurement was taken}
   \item{population}{population of the town}
}}
\source{
explained extensively on \code{inst/extdata/README.md}
}
\usage{
town_population
}
\description{
A dataset containing all of Turkey's towns populations
}
\keyword{datasets}

