% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_town_sales.R
\docType{data}
\name{town_sales}
\alias{town_sales}
\title{Town apartment sales of Turkey}
\format{A data.frame with 5 variables
\describe{
   \item{city}{name of the city}
   \item{town}{name of the town}
   \item{year}{year the measurement was taken}
   \item{month}{month the measurement was taken}
   \item{sale}{amount of the sale}
}}
\source{
explained extensively on \code{inst/extdata/README.md}
}
\usage{
town_sales
}
\description{
A dataset containing apartment sales of Turkey's towns
}
\keyword{datasets}

