% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuik_place.R
\docType{data}
\name{tuik_cities}
\alias{tuik_cities}
\title{Cities of Turkey according to tuik}
\format{An object of class \code{character} of length 81.}
\source{
city data taken from \code{\link{clean_town_population}}
}
\usage{
tuik_cities
}
\description{
Character vector of cities.
}
\keyword{datasets}

