% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuik_place.R
\docType{data}
\name{tuik_towns}
\alias{tuik_towns}
\title{Towns of Turkey according to tuik}
\format{A data.frame with 2 variables
\describe{
   \item{city}{name of the city}
   \item{town}{name of the town}
}}
\source{
town data taken from \code{\link{clean_town_population}}
}
\usage{
tuik_towns
}
\description{
Towns of Turkey according to tuik
}
\keyword{datasets}

