% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/town_sales.R
\name{get_europe_istanbul_ratio}
\alias{get_europe_istanbul_ratio}
\title{(Istanbul europe side home sales / Istanbul home sales) ratio}
\usage{
get_europe_istanbul_ratio(average = FALSE)
}
\arguments{
\item{get}{historical average, rather then returning the ratio per month.}
}
\description{
\code{town_sales} data has the granularity to calculate this ratio.
Here the calcualted ratio is the total home sales. Normally \code{town_sales}
data is not complete for each town of a city but Istanbul is the biggest
city so it is complete in this case.
}

